Move methods to a new wagtail.utils.contenttypes module
